pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "struct2token"
version = "0.1.0"
description = "Adaptive all-atom tokenization of macromolecular structures"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.3",
    "einops>=0.8",
    "biopython>=1.83",
    "scipy>=1.10",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "wandb>=0.15",
    "timm>=0.9",
    "jaxtyping>=0.2",
    "tqdm",
    "pandas>=2.0",
    "pyarrow>=12.0",
]

[project.optional-dependencies]
flash = [
    "flash-attn>=2.5",
]
dev = [
    "pytest>=7.0",
    "pytest-timeout>=2.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
